// types/pf2e/src/module/rules/synthetics.d.ts
interface RollOptionToggle {
    /** The ID of the item with a rule element for this toggle */
    itemId: string;
    label: string;
    placement: string;
    domain: string;
    option: string;
    suboptions: {
        label: string;
        selected: boolean;
    }[];
    alwaysActive: boolean;
    checked: boolean;
    enabled: boolean;
}